Revert "separating responsibility"
