--- config.json
@@ -1,5 +1,0 @@
-{
-    "modules": [
-      
-    ]
-}
